atom charge is now f64

+++ libcint/src/parse/atom.rs
@@ -36,7 +36,7 @@ pub struct AtomInfo {
     /// Element symbol without label suffix (e.g., "H", "O").
     pub symbol: String,
     /// Atomic number (Z). 0 for ghost atoms.
-    pub charge: i32,
+    pub charge: f64,
     /// Whether this is a ghost atom.
     pub is_ghost: bool,
     /// Cartesian coordinates in Bohr.
@@ -47,8 +47,11 @@ impl AtomInfo {
     /// Create a new AtomInfo from symbol and coordinates.
     pub fn new(symbol: &str, coords: [f64; 3], unit: Unit) -> Self {
         let (parsed_symbol, is_ghost) = parse_atom_symbol(symbol);
-        let charge =
-            if is_ghost { 0 } else { element_charge(&parsed_symbol).unwrap_or_else(|| panic!("Unknown element: {}", parsed_symbol)) };
+        let charge = if is_ghost {
+            0.0
+        } else {
+            element_charge(&parsed_symbol).unwrap_or_else(|| panic!("Unknown element: {}", parsed_symbol)) as f64
+        };
         let coords_bohr =
             if unit == Unit::Angstrom { [coords[0] * ANG_TO_BOHR, coords[1] * ANG_TO_BOHR, coords[2] * ANG_TO_BOHR] } else { coords };
         AtomInfo { label: symbol.to_string(), symbol: parsed_symbol, charge, is_ghost, coords: coords_bohr }
@@ -187,7 +190,8 @@ fn parse_atom_line(line: &str, default_unit: Unit) -> Result<AtomInfo, CIntError
     // Convert to Bohr if necessary
     let coords_bohr = if unit == Unit::Angstrom { [x * ANG_TO_BOHR, y * ANG_TO_BOHR, z * ANG_TO_BOHR] } else { [x, y, z] };
 
-    let charge = if is_ghost { 0 } else { element_charge(&parsed_symbol).unwrap_or_else(|| panic!("Unknown element: {}", parsed_symbol)) };
+    let charge =
+        if is_ghost { 0.0 } else { element_charge(&parsed_symbol).unwrap_or_else(|| panic!("Unknown element: {}", parsed_symbol)) as f64 };
 
     Ok(AtomInfo { label: symbol.to_string(), symbol: parsed_symbol, charge, is_ghost, coords: coords_bohr })
 }
@@ -319,8 +323,11 @@ pub fn parse_zmatrix(s: &str, unit: Unit) -> Result<Vec<AtomInfo>, CIntError> {
             },
         };
 
-        let charge =
-            if is_ghost { 0 } else { element_charge(&parsed_symbol).unwrap_or_else(|| panic!("Unknown element: {}", parsed_symbol)) };
+        let charge = if is_ghost {
+            0.0
+        } else {
+            element_charge(&parsed_symbol).unwrap_or_else(|| panic!("Unknown element: {}", parsed_symbol)) as f64
+        };
 
         atoms.push(AtomInfo { label: symbol.to_string(), symbol: parsed_symbol, charge, is_ghost, coords });
         atom_order.push(symbol.to_string());
@@ -453,11 +460,11 @@ mod tests {
         let atoms = parse_atom_string("H 0 0 0; O 0 0 1.2", Unit::Angstrom).unwrap();
         assert_eq!(atoms.len(), 2);
         assert_eq!(atoms[0].symbol, "H");
-        assert_eq!(atoms[0].charge, 1);
+        assert_eq!(atoms[0].charge, 1.0);
         assert_relative_eq!(atoms[0].coords[2], 0.0);
 
         assert_eq!(atoms[1].symbol, "O");
-        assert_eq!(atoms[1].charge, 8);
+        assert_eq!(atoms[1].charge, 8.0);
         assert_relative_eq!(atoms[1].coords[2], 1.2 * ANG_TO_BOHR);
     }
 
@@ -467,7 +474,7 @@ mod tests {
         assert_eq!(atoms.len(), 2);
         assert!(!atoms[0].is_ghost);
         assert!(atoms[1].is_ghost);
-        assert_eq!(atoms[1].charge, 0);
+        assert_eq!(atoms[1].charge, 0.0);
     }
 
     #[test]
@@ -476,31 +483,31 @@ mod tests {
         let atoms1 = parse_atom_string("GHOST-O 0 0 1.2", Unit::Angstrom).unwrap();
         assert_eq!(atoms1.len(), 1);
         assert!(atoms1[0].is_ghost);
-        assert_eq!(atoms1[0].charge, 0);
+        assert_eq!(atoms1[0].charge, 0.0);
 
         // Test ghost atom with real atom (two atoms)
         let atoms2 = parse_atom_string("GHOST-O 0 0 1.2; O 0 0 0", Unit::Angstrom).unwrap();
         assert_eq!(atoms2.len(), 2);
         assert!(atoms2[0].is_ghost);
         assert!(!atoms2[1].is_ghost);
-        assert_eq!(atoms2[0].charge, 0);
-        assert_eq!(atoms2[1].charge, 8);
+        assert_eq!(atoms2[0].charge, 0.0);
+        assert_eq!(atoms2[1].charge, 8.0);
 
         // Test X_H format
         let atoms3 = parse_atom_string("X_H 0 0 1; H 0 0 0", Unit::Angstrom).unwrap();
         assert_eq!(atoms3.len(), 2);
         assert!(atoms3[0].is_ghost);
         assert!(!atoms3[1].is_ghost);
-        assert_eq!(atoms3[0].charge, 0);
-        assert_eq!(atoms3[1].charge, 1);
+        assert_eq!(atoms3[0].charge, 0.0);
+        assert_eq!(atoms3[1].charge, 1.0);
 
         // Test ghost.H format (lowercase with dot separator)
         let atoms4 = parse_atom_string("ghost.H 0 0 1; H 0 0 0", Unit::Angstrom).unwrap();
         assert_eq!(atoms4.len(), 2);
         assert!(atoms4[0].is_ghost);
         assert!(!atoms4[1].is_ghost);
-        assert_eq!(atoms4[0].charge, 0);
-        assert_eq!(atoms4[1].charge, 1);
+        assert_eq!(atoms4[0].charge, 0.0);
+        assert_eq!(atoms4[1].charge, 1.0);
 
         // Test multiple ghost atoms
         let atoms5 = parse_atom_string("GHOST-O 0 0 0; X_H 0 0 1; ghost.H 0 0 2; O 0 0 3; H 0 0 4; H 0 0 5", Unit::Angstrom).unwrap();
@@ -519,7 +526,7 @@ mod tests {
         let atoms = parse_atom_string("ghost.H 0 0 1", Unit::Angstrom).unwrap();
         assert_eq!(atoms.len(), 1);
         assert!(atoms[0].is_ghost);
-        assert_eq!(atoms[0].charge, 0);
+        assert_eq!(atoms[0].charge, 0.0);
         assert_eq!(atoms[0].symbol, "H");
     }
